requirements.txt
Django>=4.0.3,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
django-extensions>=3.1.5,<3.2
djangorestframework-simplejwt>=5.1.0,<5.2